+++ src/bin/llhd-opt.rs
@@ -25,15 +25,12 @@ fn main() {
 }
 
 fn main_inner() -> Result<(), String> {
+    // Configure the logger.
+    pretty_env_logger::init_custom_env("LLHD_LOG");
+
+    // Parse the command line arguments.
     let matches = app_from_crate!()
         .about("Optimizes LLHD assembly.")
-        .arg(
-            Arg::with_name("verbosity")
-                .short("v")
-                .multiple(true)
-                .help(HELP_VERBOSITY.lines().next().unwrap())
-                .long_help(HELP_VERBOSITY),
-        )
         .arg(
             Arg::with_name("input")
                 .help("LLHD file to optimize")
@@ -76,15 +73,6 @@ fn main_inner() -> Result<(), String> {
                 .help("Execute passes to lower behavioural to structural LLHD"),
         )
         .get_matches();
-
-    // Configure the logger.
-    let verbose = matches.occurrences_of("verbosity") as usize + 1;
-    stderrlog::new()
-        .module("llhd")
-        .module("llhd_opt")
-        .verbosity(verbose)
-        .init()
-        .unwrap();
 
     // Configure rayon to be single-threaded if requested.
     if matches.is_present("single-threaded") {
@@ -239,16 +227,6 @@ fn main_inner() -> Result<(), String> {
         Ok(())
     }
 }
-
-static HELP_VERBOSITY: &str = "Increase message verbosity
-
-This option can be specified multiple times to increase the level of verbosity \
-in the output:
-
--v    Print info messages
--vv   Also print debug messages
--vvv  Also print detailed tracing messages
-";
 
 static HELP_PASSES: &str = "Exact order of passes to run